fix typescript usage
